refactor(works): 関連作品の Firestore アクセスを action に切り出す (SPR-302)
AI レビュー指摘。リポジトリ内で .tsx から直接 getFirestore() を呼んでいるのは
work-related-works.tsx だけで、第1弾の work-price-summary.tsx を含め他は全て
.ts 側に置かれていた。コンポーネントはキャッシュ境界と JSX の組み立てだけを持つ形に揃える。

共有層（src/actions/）ではなく route 同居にしたのは、fetchWorksByIds と
pickRelatedWorks が works route 配下にあり、共有層から route 層への逆依存が
禁止されているため（同ディレクトリの evaluation-actions.ts と同じ置き方）。

出力は不変。Emulator で JSON-LD（Product / offers / aggregateRating / contributor 4名）と
関連作品リンク4件が変更前と同一であることを確認済み。
