+++ apps/web/src/app/works/[workId]/components/work-related-works.tsx
@@ -1,11 +1,8 @@
-import type { CircleDocument } from "@suzumina.click/shared-types";
 import { Card, CardContent, CardHeader, CardTitle } from "@suzumina.click/ui/components/ui/card";
 import { cacheLife } from "next/cache";
 import Link from "next/link";
-import { getFirestore } from "@/lib/firestore";
-import { fetchWorksByIds } from "../../utils/fetch-works-by-ids";
-import { convertWorksToPlainObjects } from "../../utils/work-converters";
-import { pickRelatedWorks, type RelatedWork } from "../related-works";
+import type { RelatedWork } from "../related-works";
+import { getCircleRelatedWorks } from "../related-works-actions";
 
 /**
  * 同じサークルの他の作品（SPR-302 第2弾）。**Server Component であることが要件**。
@@ -18,27 +15,14 @@ import { pickRelatedWorks, type RelatedWork } from "../related-works";
  *
  * 読み取りコスト: circles/{id} が 1 read ＋ 表示分の works が最大 RELATED_WORKS_LIMIT read。
  * Firestore reads は課金上位のため `use cache` + `cacheLife("days")` で 1 作品 1 日 1 回に抑える。
- * **workIds 全件を引かない**ことが重要（789作品のサークルが存在し、全件取得は #917 で
- * 潰したのと同じ全件スキャンになる）。
+ * Firestore アクセス自体は `related-works-actions.ts` に置く（第1弾の `work-price-summary.tsx`
+ * と同じく、コンポーネントはキャッシュ境界と JSX の組み立てだけを持つ）。
  */
 async function loadRelatedWorks(circleId: string, currentWorkId: string): Promise<RelatedWork[]> {
 	"use cache";
 	cacheLife("days");
 
-	const firestore = getFirestore();
-	const circleDoc = await firestore.collection("circles").doc(circleId).get();
-	if (!circleDoc.exists) return [];
-
-	const circleData = circleDoc.data() as CircleDocument;
-	const targetIds = pickRelatedWorks(circleData.workIds ?? [], currentWorkId);
-	if (targetIds.length === 0) return [];
-
-	const works = await fetchWorksByIds(firestore, targetIds);
-	return convertWorksToPlainObjects(works).map((work) => ({
-		productId: work.productId,
-		title: work.title,
-		formattedPrice: work.price.formattedPrice,
-	}));
+	return getCircleRelatedWorks(circleId, currentWorkId);
 }
 
 export default async function WorkRelatedWorks({
